+++ racingline/src/main/java/org.avphs.racingline/org/avphs/racingline/RacingLine.java
@@ -1,0 +1,35 @@
+package org.avphs.racingline;
+
+import java.util.ArrayList;
+import java.util.HashSet;
+
+/**
+ * <p>This class represents a racing line. It contains an array of points which represent the line.</p>
+ *
+ * @see RacingLinePoint
+ */
+public class RacingLine {
+    private ArrayList<RacingLinePoint> RacingLinePointsList = new ArrayList<RacingLinePoint>();
+    private RacingLinePoint[] RacingLinePoints;
+
+    public RacingLine() {
+
+    }
+
+    public RacingLinePoint[] getRacingLinePoints() {
+        return RacingLinePoints;
+    }
+
+    public void addPoint(RacingLinePoint newPoint) {
+        RacingLinePointsList.add(newPoint);
+    }
+
+    public void sortPoints() {
+        ArrayList<RacingLinePoint> orderedRacingLine = new ArrayList<RacingLinePoint>();
+        HashSet<RacingLinePoint> remainingPoints = new HashSet<RacingLinePoint>(RacingLinePointsList);
+        orderedRacingLine.add(RacingLinePointsList.get(0));
+        remainingPoints.remove(RacingLinePointsList.get(0));
+        while(remainingPoints.size() > 0){
+        }
+    }
+}
