+++ racingline/src/main/java/org.avphs.racingline/org/avphs/racingline/RacingLineModule.java
@@ -15,8 +15,8 @@ public class RacingLineModule implements CarModule {
     private boolean[][] map;
     private boolean[][] visited;
     private boolean[][] added;
-    boolean addToOuter;
-    int length, width;
+    private boolean addToOuter;
+    private int length, width;
     private int[] dx = {-1, 0, 1, 0};
     private int[] dy = {0, 1, 0, -1};
 
@@ -43,7 +43,6 @@ public class RacingLineModule implements CarModule {
     //endregion
 
     //region RacingLine
-
     /**
      * This method creates the racing line. This should be run before getRacingLine is called.
      *
@@ -145,100 +144,6 @@ public class RacingLineModule implements CarModule {
 }
 
 //region Classes
-
-/**
- * <p>This class represents a racing line. It contains an array of points which represent the line.</p>
- *
- * @see RacingLinePoint
- */
-class RacingLine {
-    private ArrayList<RacingLinePoint> RacingLinePointsList = new ArrayList<RacingLinePoint>();
-    private RacingLinePoint[] RacingLinePoints;
-
-    public RacingLine() {
-
-    }
-
-    public RacingLinePoint[] getRacingLinePoints() {
-        return RacingLinePoints;
-    }
-
-    public void addPoint(RacingLinePoint newPoint) {
-        RacingLinePointsList.add(newPoint);
-    }
-
-    public void sortPoints() {
-        ArrayList<RacingLinePoint> orderedRacingLine = new ArrayList<RacingLinePoint>();
-        HashSet<RacingLinePoint> remainingPoints = new HashSet<RacingLinePoint>(RacingLinePointsList);
-        orderedRacingLine.add(RacingLinePointsList.get(0));
-        remainingPoints.remove(RacingLinePointsList.get(0));
-        while(remainingPoints.size() > 0){
-
-        }
-    }
-}
-
-class RacingLinePoint {
-    private float x, y, degree;
-
-    public RacingLinePoint() {
-        setX(0);
-        setY(0);
-        setDegree(0);
-    }
-
-    public RacingLinePoint(float x, float y) {
-        setX(x);
-        setY(y);
-        setDegree(0);
-    }
-
-    public RacingLinePoint(float x, float y, float degree) {
-        setX(x);
-        setY(y);
-        setDegree(degree);
-    }
-
-    public float getX() {
-        return x;
-    }
-
-    public float getY() {
-        return y;
-    }
-
-    public float getDegree() {
-        return degree;
-    }
-
-    public void setX(float x) {
-        this.x = x;
-    }
-
-    public void setY(float y) {
-        this.y = y;
-    }
-
-    public void setDegree(float degree) {
-        this.degree = degree;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        RacingLinePoint that = (RacingLinePoint) o;
-        return Float.compare(that.x, x) == 0 &&
-                Float.compare(that.y, y) == 0;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(x, y);
-    }
-
-}
-
 class WallPoint {
     int x, y;
