+++ racingline/src/main/java/org.avphs.racingline/org/avphs/racingline/RacingLinePoint.java
@@ -1,0 +1,63 @@
+package org.avphs.racingline;
+
+import java.util.Objects;
+
+public class RacingLinePoint {
+    private float x, y, degree;
+
+    public RacingLinePoint() {
+        setX(0);
+        setY(0);
+        setDegree(0);
+    }
+
+    public RacingLinePoint(float x, float y) {
+        setX(x);
+        setY(y);
+        setDegree(0);
+    }
+
+    public RacingLinePoint(float x, float y, float degree) {
+        setX(x);
+        setY(y);
+        setDegree(degree);
+    }
+
+    public float getX() {
+        return x;
+    }
+
+    public float getY() {
+        return y;
+    }
+
+    public float getDegree() {
+        return degree;
+    }
+
+    public void setX(float x) {
+        this.x = x;
+    }
+
+    public void setY(float y) {
+        this.y = y;
+    }
+
+    public void setDegree(float degree) {
+        this.degree = degree;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        RacingLinePoint that = (RacingLinePoint) o;
+        return Float.compare(that.x, x) == 0 &&
+                Float.compare(that.y, y) == 0;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(x, y);
+    }
+}
